refactor: rename SplashScreen section to Hero
